requirement.txt
Flask
mysql-connector-python
werkzeug